use horovod

+++ dqn_learner.py
@@ -8,9 +8,13 @@ from collections import deque
 
 import gym
 import numpy as np
+import tensorflow as tf
 from tensorflow.keras.layers import Dense
 from tensorflow.keras.models import Sequential
 from tensorflow.keras.optimizers import Adam
+
+import horovod.tensorflow.keras as hvd
+from tensorflow.keras import backend as K
 
 
 class DQNAgent:
@@ -32,7 +36,7 @@ class DQNAgent:
         model.add(Dense(24, input_dim=self.state_size, activation='relu'))
         model.add(Dense(24, activation='relu'))
         model.add(Dense(self.action_size, activation='linear'))
-        model.compile(loss='mse', optimizer=Adam(lr=self.learning_rate))
+        model.compile(loss='mse', optimizer=hvd.DistributedOptimizer(Adam(lr=self.learning_rate * hvd.size())))
         return model
 
     def memorize(self, state, action, reward, next_state, done):
@@ -62,14 +66,22 @@ if __name__ == '__main__':
     state_size = env.observation_space.shape[0]
     action_size = env.action_space.n
 
+    hvd.init()
+    config = tf.ConfigProto()
+    config.gpu_options.allow_growth = True
+    config.gpu_options.visible_device_list = str(hvd.local_rank())
+    K.set_session(tf.Session(config=config))
+    hvd.callbacks.BroadcastGlobalVariablesCallback(0)
+
     agent = DQNAgent(state_size, action_size)
 
     os.environ['KMP_WARNINGS'] = '0'
+    os.environ["TF_CPP_MIN_LOG_LEVEL"]='1'
     socket = zmq.Context().socket(zmq.ROUTER)
     socket.bind("tcp://*:6080")
 
     model_path = "./model_weights"
-    if os.path.exists(model_path):
+    if hvd.rank() == 0 and os.path.exists(model_path):
         os.system("rm -rf " + model_path)
     os.mkdir(model_path)
 
@@ -100,7 +112,7 @@ if __name__ == '__main__':
                 # with open(model_path + "/syf-eposide_{}-time_{}.h5".format(e, time), "rb") as file:
                 #     socket.send_multipart([id, file.read()])
 
-        if e % 3 == 0:
+        if e % 3 == 0 and hvd.rank() == 0:
             print("learner -- send model")
             agent.save(model_path + "/syf-eposide_{}.h5".format(e, time))
             with open(model_path + "/syf-eposide_{}.h5".format(e, time), "rb") as file:
